Rename compare-event-ordering
